+++ packages/react/src/components/Modal/styles.ts
@@ -50,7 +50,7 @@ export const ModalContent = styled(Content, {
   left: '50%',
   transform: 'translate(-50%, -50%)',
   width: '100%',
-  height: '100%',
+  height: '100% !important',
   zIndex: '$max',
 
   '@media (prefers-reduced-motion: no-preference)': {