+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @compasso/design-tokens
 
+## 2.1.0
+
+### Minor Changes
+
+- General components UI updates
+
 ## 2.0.0
 
 ### Major Changes